requirements.txt
Django==2.1.5
django-polymorphic==2.0.3
psycopg2-binary==2.7.7
pytz==2018.9
